Add ABSL_FLAG -producer_side_server to OrbitService
The Unix socket on which the producer-side server (which receives data from
Vulkan layer, manual instrumentation, user space instrumentation) listens on is
used to determine whether another instance of OrbitService is already running.

The new `-producer_side_server` flag allows to not start another producer-side
server and to skip this verification when starting another instance of
OrbitService (using a different TCP port for the communication with Orbit UI)
with the purpose of profiling the first instance of OrbitService.

In the end, I decided to go for a "positive" version of the flag with a `true`
default (i.e., `-producer_side_server` instead of `-no_producer_side_server`)
because `bool` Abseil flags already allow a "no" prefix to specify `false`
(i.e., `-noproducer_side_server` is equivalent to
`-producer_side_server=false`).

Test: Take a capture of a "regular" instance of OrbitService with an instance of
OrbitService that uses `-noproducer_side_server` and `-grpc_port`.