Fix function list update (#3862)
This triggers `FireRefreshCallback` always after
`UpdateAfterSymbolLoading`, instead of inside
`UpdateAfterSymbolLoading`. The call inside of
`UpdateAfterSymbolLoading` happened at the end and only in the case
where a capture data was available. This meant that the function list
was not updated before there was a capture.

Compare: https://github.com/google/orbit/pull/3825/

Test: Start Orbit, let symbols load, function list is updated.
Bug: http://b/237503865