Implement logic for the list of threads in Mizar GUI
The list should show the list of threds we have samples for
in the descending order (that is, just like under the Sampling
tab of Orbit).

Multiple selection is allowed. When selection changes, the
state of SamplingWithFrameTrackInputWidget is updated.

For the sake of testability `SamplingWithFrameTrackInputWidget`
is templatized now. `Q_OBJECT` cannot be templatized, but its base
can, so `SamplingWithFrameTrackInputWidgetBase` was introduced.

Also, some getters are added to `BaselineAndComparison`.

`MizarPairedData` cllects some statistics upon creation now.

Tests: Compile, Unit, Manual
Bug: http://b/237532309