Only perform epilog detection when unwinding the innermost frame
If we are not in the innermost frame, we know we cannot have executed any
instructions of an epilog, because the PC is right after a call instruction, and
call instructions cannot appear in epilogs. The PC could still be on the very
first instruction of an epilog, but here unwinding by processing all the
`UNWIND_CODE`s (just like when normally outside of the prolog and the epilogs)
is equivalent.

This is a big performance optimization: for example, detecting an epilog
requires disassemblying, which is expensive. In our testing, average time of
unwinding one frame is approximately halved.

Also add somewhat unrelated `step_fails_if_eval_fails_inside_of_prolog` test for
completeness.

Bug: http://b/237625163

Test:
- Add unit test.
- Capture a complex game running on Wine and verify no worsening of unwind error
  rate.
- Verify performance gain with ad-hoc instrumentation.