Make PeCoffUnwindInfos::GetUnwindInfo return a pointer to an UnwindInfo
This makes `PeCoffUnwindInfos::GetUnwindInfo` avoid a copy when returning an
`UnwindInfo`. Since this is a complex type, containing in particular a
`std::vector<UnwindCode>`, this is a measurable optimization. In my test, this
makes unwinding a single frame around 10% faster.

Note that `PeCoffUnwindInfosImpl` returns a pointer to an `UnwindInfo` contained
in an `std::unordered_map`. For this container, "references and pointers to
either key or data stored in the container are only invalidated by erasing that
element, even when the corresponding iterator is invalidated", so the returned
pointer stays valid even if another call to `GetUnwindInfo` causes the container
to be rehashed.

Bug: http://b/237625163

Test: Capture complex game running on Wine.