Return from PeCoff::Step if object is invalid
When we can't initialize the PeCoffInterface member of PeCoff, the
PeCoff instance is marked as invalid and the interface member is reset
to a nullptr. This leads to a crash when trying to carry out a step in
such a case. Analogously to the Elf case, we return from PeCoff::Step
if the object is marked as invalid. Adjusted tests to match this
changed behavior.

Tested: Unit tests; tested on game.
Bug: http://b/237649102