Always use the executor in ~ScopedStatus
At some point I made ScopedStatus keep a weak pointer to an executor to
avoid any threading issues. This allows ScopedStatus to perform
operations on the main thread even when the object lives in a different
thread. This is often the case because ScopedStatus objects are useful
in particular for background operations.

A similar issue can occur entirely on the main thread when the
StatusListener outlives the ScopedStatus object. But since the Executor
is supposed to always be destroyed before the StatusListener we can
avoid the situation by always accessing the StatusListener through the
executor.

Bug: http://b/237746722