Cache addresses for which not an epilog was successfully detected
While we cannot cache the addresses for which we detected an epilog, as we
actually have to simulate the effects of the rest of that epilog, we can cache
the addresses for which we detected that we are _not_ in an epilog. These are
the majority anyway, as most samples are not going to be in an epilog.

In my usual experiments, this is a 15 to 20% performance improvement on the
average time to unwind one frame, for example because it saves us from a lot of
disassemblying.

Bug: http://b/237625163

Test:
- Add unit test.
- Capture a complex game running on Wine and verify no worsening of unwind error
  rate.
- Verify performance gain with ad-hoc instrumentation.