Add expansion slider to TopDownView in devmode (#3770)
Add a slider in CallTreeWidget to quickly control the expansion "level"
of the nodes. The slider fully to the left is replicating the current state,
which is to have everything collapsed by default. The slider all the way
to the right makes the tree fully expand. The values in between expand
nodes progressively based on their "Inclusive" percentage.

The feature is in "devmode" only for now for internal validation.

https://screenshot.googleplex.com/Brm6xsZ6gfex8QR

b/235482436