Avoid sending version comparison warning to UI
We changed to stop sending user the warning message that the capture is
taken by a newer version.

When we load a capture, we already compare the capture format version in
`OrbitApp::LoadCaptureFromFile` (to be specific, when we call
CaptureFile::OpenForReadWrite.
This should be enough for comparing the version compatibility.

Bug: http://b/237747877