Adapt connect_to_target test to use instance ID
The test can't be executed by the CI as it is written currently since
the go-code does not have access to the instance display name, but this
is the only information displayed in the main window to verify the
correct instance.

With this change, functionality to read tooltips is added, and the test
uses the target_label tooltip to check the contained instance id and
process path.