Improve frozen UI on capture loading (#3887)
This improves an issue with `update_after_symbol_loading_throttle_` in
OrbitApp. The throttle did not work when the call to
`UpdateAfterSymbolLoading()` (which does post processing) took longer
than the throttle interval (in this case more than 1 second). This is
fixed by using a `Qt::QueuedConnection`, to queue the post processing
instead of executing it directly.

Test: Manual
Bug: http://b/237983515