Remove ModuleSymbols's load_bias and symbols_file_path
These were only used in two places, because everywhere else we use the
information from `ModuleInfo` and `ModuleData`. In fact, they confusingly
duplicate the information that is in `ModuleInfo`, but only partly. Also,
`load_bias` with the purpose of translating from `SymbolInfo::address` to the
offset in the file is only correct for ELF files.

Bug: http://b/235824531

Test: Build.