Open uprobes for functions to collect stack for
This adds a new proto message to send the functions to record
the stack on to the service.
In the service we open uprobes for those functions and
read the perf events. So far, the empty default
implementation of the visitor will handle
those events.

This also unifies the "forwarding to the ringbuffer"
for uprobes and the implementation for all our
tracepoints (b/238164435).

Test: Manual E2E prototype.
Bug: http://b/236118579
Bug: http://b/238164435