Add "Wine Syscall Handling Method" to the client
This adds an enum "WineSyscallHandlingMethod" to the client to
specify how to deal with the wine syscalls.
Depending on the value set in DataManager, App will communicate
to the client to stop unwinding at the dispatcher or to
record the user stack at the dispatcher.

The option is not yet exposed to the user and the service
will not yet handle the case to record the user stack.

Test: Manual E2E prototype
Bug: http://b/236118579