Add stack slice class
This adds a thin layer to store and process some stack data (collected
by perf_event_open) in memory. It comes with a `StackSlice` class
meant to store the stack copy in memory and a `StackSliceView` class
meant to pass the stack slice to functions and avoid copies.

Note: The main reason to not just pass pointers/references of
`StackSlice` into functions, is that this allows us to store
the `StackSliceView` in tests and validate its content even
after the `StackSlice` has ended its lifetime.

Test: Run E2E test
Bug: http://b/236118579