Split LinuxMap into ReadMaps and ReadModules
`ParseMaps` was performing two tasks inside the same functions: list the maps
from `/proc/[pid]/maps`, and list the modules from the list of maps. We split
this in two.
The more practical reason for this change is that I will need the intermediate
result (list of maps) in order to detect whether a function can be instrumented
by uprobes (i.e., whether a file mapping with the function's module and file
offset exists).

Bug: http://b/232072696

Test:
- Unit tests.
- Start Orbit on `triangle.exe`. Verify modules show up correctly.