Update libunwindstack to c4e572564862727056090d9540641077238f2079 (#3937)
* Use #pragma once header guard.

Move to using the #pragma once header guard instead of the #ifndef
construct since it's cleaner.

Test: Compiles.
Change-Id: I0bd76b53cfc78df7465029ef34783f85651a9c74
(cherry picked from commit 0cd3943fb8857eba64f8590fd6137648f3fcb198)

* Use #pragma once also in PE-related files

* Promote HWASan tests to presubmit.

They have been passing and do not show significant slowdown.

Change-Id: I29808d704ae906fcc901c1c6201d7608b36298f6
(cherry picked from commit c6675b9a36d7a00d11a4555a065f9431e4440656)

* Fix DW_OP_bra handling.

The handling of the DW_OP_bra is incorrect. The spec states:

DW_OP_bra is a conditional branch. Its single operand is a 2-byte
signed integer constant. This operation pops the top of stack. If
the value popped is not the constant 0, the 2-byte constant
operand is the number of bytes of the DWARF expression to skip
forward or backward from the current operation, beginning after
the 2-byte constant.

The problem is the code still used the operand even if the constant
is zero.

Modified the unit test to verify that the operand is ignored
if the stack value is zero.

Bug: 237751512

Test: Unit tests pass.
Change-Id: I012fda541b40ea65a381963c589ad7c0334e88f3
(cherry picked from commit c4e572564862727056090d9540641077238f2079)

Co-authored-by: Christopher Ferris <cferris@google.com>
Co-authored-by: Florian Mayer <fmayer@google.com>