Fix use of wrong vector in RetrieveInstancesWidget
This does not actually change any logic, but only improves readability.
It was a mistake before that has now been fixed. For explanation: The
variable `projects` is a reference to `initial_load_result.projects` and
projects is then sorted. Since it is a reference, this actually changes
`initial_load_result` and the code worked before, even when
`initial_load_result` was used after the call to sort. To make this less
confusing, now `projects` is used after the call to sort.