Add & use ComputeVirtualAddressOfInstrumentedFunctionsIfNecessary
Follow up from https://github.com/google/orbit/pull/3918: instead of checking
whether `InstrumentedFunction::function_virtual_address()` is available in
multiple parts of the code, simply compute it from `file_offset()` right at the
end of the capture if it is not available (in `OrbitApp::OnCaptureComplete`,
where we do other post-processing).

This allows for further clean-ups.

Bug: http://b/235824531

Test:
- Take captures with instrumented functions, both with uprobes and user space
  instrumentation.
- Load an older capture with instrumented functions. Enable and disable frame
  tracks.