Remove some functions using file offsets and not addresses, fix others (#3918)
We want to identify functions by virtual address now, not by offset in the file
(whose computation from the virtual address is more complex and different
between ELF files and PEs).

We remove `SymbolOffsetToAbsoluteAddress` and `SymbolAbsoluteAddressToOffset`,
in favor of `SymbolVirtualAddressToAbsoluteAddress` and
`SymbolAbsoluteAddressToVirtualAddress`.

We also remove `ModuleData::FindFunctionByOffset` which is now unused.

We need to add `InstrumentedFunction::function_virtual_address()`, and we use it
where necessary. Note that capture compatibility still requires falling back to
`InstrumentedFunction::function_offset()` in some places.

The remaining functions dealing with offsets are `FunctionInfo::FileOffset`
(also renamed) and `FindFunctionByModulePathBuildIdAndOffset`. We fix those, by
using `ModuleData::ConvertFromVirtualAddressToOffsetInFile` and
`ModuleData::ConvertFromOffsetInFileToVirtualAddress`.

Bug: http://b/235824531

Test:
- Take captures with instrumented functions, both with uprobes and user space
  instrumentation.
- Load an older capture with instrumented functions. Enable and disable frame
  tracks.