Automatically add a Frame Track for QueuePresentKHR if available
In this PR we are adding the logic for automatically adding a FrameTrack
for QueuePresentKHR (ggpvlk.so) if the module is available.

We are including this method as part of LoadAllSymbols, as the plan is
adding a FrameTrack as soon as the symbols for the needed module are
loaded without waiting for all of them. It's also behind a flag that now is
false and will be turned on after branch but (http://b/239149548).

Tests: Locally, set the flag to true and create some preset files locally:
 - Check that the frame-track is added if the module is available.
 - Check nothing happened when the module is not available to be loaded.
 - Check nothing happened when the preset file doesn't exist.
 - Take a capture before the ggpvlk module is loaded, check there is no
   FrameTrack. Stop the capture, wait for the module to be loaded. Take
   another capture. Check the FrameTrack was added.
 - Add another preset file to the list, check the second preset file is
   added only when the first one is not loadable.

Bug: http://b/239148765.

Next steps:
- http://b/239148767 (Adapt E2E tests).
- http://b/239155921 (Ship the default frame track as a file - after
  branch cut).