Ignore all target flags when re-instantiating Orbit
When opening a capture file from an Orbit instance that was connected
using --taret_process / --target_instance or --target_uri, loading that
file fails. Orbit starts a new instance and copies all command line
parameters, but the aforementioned parameters should *not* be passed to
the new instance. This was broken when changing the parameter name.

Bug: b/238987946
Test: Manual testing - specify --target_process and --target_instance,
then open a capture