Use "low-frequency" IDs for PresentEvent (#3961)
This event is used to record swap-chain presentation events on Windows. As such,
its frequency is the frame rate, so it's much lower than what we would like to
use "high-frequency ID"s for. Switch to "low-frequency ID"s before the IDs are
set in stone by the 1.82 branch cut.

Context is https://github.com/google/orbit/pull/3934.

Test: Build.