Wait for dataviews to be updated while checking symbols (#3982)
In this PR we are fixing the flakyness in coff-e2e test for the 1.82
release. The test is flaky because function_data_view has a lot of
functions because auto-symbol-loading and it's not reacting instantly
to modifications of the filter. We had a similar issue fixed in
https://github.com/google/orbit/pull/3860/.

In addition we are fixing a wrong condition (> 0 instead of >1) and I
formatted that file using yapf. The real change is only in
VerifySymbolLoaded.

Bug: http://b/239511821
Test: Not yet. I will create a signed build and run the E2E tests on
that PR.