Do not expose absolute addresses in ModuleData
ModuleData exists in ModuleManager once per path and build id.
So if a module is loaded twice at two different locations
in memory, there will be only one instance of ModuleData.

Thus, it must not expose any information about where it
is loaded in memory. For this, ProcessData should be
used.

Test: Check manaual instrumentation still works on OrbitTest
Bug: http://b/236710308