Link vulkan layer with hidden symbols
This changes the build of our Vulkan layer such that the shared library
only exposes the symbols that are actually needed by the Vulkan loader.

The `VK_LAYER_EXPORT` macro already takes care of marking the relevant
functions as non-hidden.

This reduces the size of the library from 6.9 MiB to 5.5 MiB.