Send target process state after capture as a metric.
Added data to ProducerCaptureEvent. The event is produced in
LinuxCaptureServiceBase and handed over to the ProducerEventProcessor.

ProducerEventProcessor merely transfers the event to a ClientCaptureEvent.

The event is finally recieved in App.cpp and added to CaptureMetric
where it is converted to OrbitCaptureData which is the proto containing the
metrics.

Test: Added logging in CaptureMetric::SendCaptureSucceeded. The correct
data ends up in the metrics proto.
Bug: b/237655477