Minimize calls to ReadModules in UserSpaceInstrumentation (#3993)
`ReadModules` is an expensive operation: it reads and parses the maps, it opens
and parses object files, and it prints plenty of information to the logs (which
made this very visible). On `triangle.exe`, it takes ~8 ms.
Previously, we were calling it in every function that needed the modules. Also,
the number of calls was growing with the number of instrumented modules.
Now we try to call it as few times as possible and pass the result down. Of
course this adds a parameter to various functions.

Test: Capture with user space instrumentation and manual instrumentation. Verify
they work and verify from the logs that the number of calls to `ReadModules` is
now reduced.