Add orbit_uprobes_wine.py E2E test script
This test verifies uprobes on `triangle.exe` running on Wine. It instruments two
functions, one from a Wine PE that has file alignment compatible with `mmap`'s
file mappings, and one from `triangle.exe`, which doesn't have file alignment
compatible with `mmap`'s file mappings.
We expect the first function to be recorded and the second one to have no hits,
but with a corresponding warning showing up in the Capture Log.

Bug: http://b/232072696

Test: Ran it locally.