Integrate ETW frame tracks into Mizar
This introduces (in FrameTrack.h) types allowing for the
uniforme handling of TimerInfos and PresentEventss.

The part responsible for trame track data handling is
abstracted out from MizarPairedData to FrameTrackManager.

ETW frame tracks are now shown on the GUI and can be selected
in the same way D and MS can.

Tets: Manual, Unit
Bug: http://b/239937976