Move ReadMaps from LinuxTracingUtils to ReadLinuxMaps (#4000)
This now seems to be the perfect place for it. Also add a smoke test and make it
return `ErrorMessageOr`.

We have to do some renaming as a consequence to prevent confusion.

The previous `ReadMaps(std::string_view)` becomes `ParseMaps`, also in line with
`LibunwindstackMaps::ParseMaps`. `ReadMaps(pid_t)` becomes `ReadAndParseMaps`.

As with "parse maps" we now mean from text to `LinuxMemoryMapping`s, we change
`ParseMapsIntoModules` to `ReadModulesFromMaps` (in line and contrast with
`ReadModules`, that takes the pid).

Test: Capture `triangle.exe`. Verify modules show up correctly, callstacks make
sense.