Add top-down/bottom-up inspections (#3990)
When a user chooses to "Select these callstacks" in the context menu in
top-down/bottom-up tabs, this creates the selection within the tab
itself.  It also adds a banner above to show the user that they are
currently in an inspection and give them the option to leave the
inspection.

The effect on the samples tab will be done in an upcoming PR.

These changes are hidden behind the time_range_selection flag.