Add auto frame track at most once per session (#4016)
In this PR we are addressing the issue of a FrameTrack being re-added
after being manually erased by the user. The FrameTrack will appear
again in the next session unless the user unselect the option in
user-settings.

Discussed in http://go/stadia-orbit-auto-frame-track-alignment.

Bug: http://b/239675491

Test:
1- Start a session, take a capture, auto frame track is there.
2- Manually erase it, check Frame Track is not there anymore in the
following captures.

PS: It could be nice to have an E2E tests that plays both with the auto
frame track user setting (turning it on and off) and erasing manually
the FrameTrack, but IMO it will mean adding a new E2E tests and I'm not
sure if its a priority. Let me know if you have a different opinion.