Test the new auto frame track setting in E2E tests (#4003)
* Test the new auto frame track setting in E2E tests

In this PR we are modifying 4 E2E tests to test the new Auto Frame Track
user settings. Two of them are setting the setting to true, and two of
them to false. As this will change the addition of a FrameTrack in the
next session, we are ending the session and starting another one.

Basically I created a new class but most of the used functions were
written in a shared utility class. There is a TODO here, since I didn't
refactor SetAndCheckMemorySamplingPeriod to use these shared methods,
but this will be made in a following PR.

This PR won't be merged before
https://github.com/google/orbit/pull/4001.

Bug: http://b/239148938.

Test: Manually run E2E tests. Created a signed build and run E2E test
for https://github.com/google/orbit/pull/3998.

* Fixup: Explanation about why it's not a paramenter in Capture