Prefer COFF symbol table to DWARF for CoffFileImpl::LoadDebugSymbols
Previously, we were extracting symbols from PEs using DWARF unwind information,
present in Wine DLLs. However, many symbols are missing from this, including
*but not limited to* leaf functions, so we had to integrate this information
with the symbols from the COFF symbol table. However: symbols from the COFF
symbol table don't have sizes, which we had to deduce as the distance to the
next symbol; some functions from the two sources were differing by one
eight-byte instruction, which caused those functions to show up twice with two
different addresses.

We now turn this around and prefer loading symbols from the COFF symbol table,
as it's much more complete. For non-leaf functions, we get the sizes from the
`RUNTIME_FUNCTION`s, for which we add `CoffFileImpl::GetUnwindRanges`. In
general, this already gives more functions with the exact size set than before
(again, even excluding leaf functions, in general we have more functions in the
COFF symbol table with a `RUNTIME_FUNCTION` than in the DWARF information).
There are still rare cases of functions only present in the DWARF unwind
information: we still add these, but afterwards, and only considering address
ranges that don't intersect with any symbol from the COFF symbol table, in order
to avoid the problem mentioned above.

Test: Apart from updating the unit test, correctness of the approach was
verified by collecting various information on the differences between the COFF
symbol table and the DWARF unwind information for all the Wine DLLs; this
included in particular the problematic functions mentioned above, i.e.,
functions with a different start address in the two sources, and functions only
present in the DWARF information. After the change, I verified for each of those
DLLs that those functions are handled correctly.

Note that `CoffFileImpl::GetUnwindRanges` is not unit tested in isolation, but
as part of `LoadDebugSymbols`. It will be further unit tested as part of using
those address ranges as fallback symbols.