Verify that pdata bounds are valid for the underlying file
The bounds for the pdata section are read from the PE/COFF file itself,
which means they are untrusted data. We add a check to verify that the
bounds fall within the underlying PE/COFF file. This also ensures that
we do not allocate absurdly large memory regions to hold the
RUNTIME_FUNCTIONS structs, which led to an OOM issue found by oss-fuzz.

Tested: Unit tests; ran fuzzer testcase to verify OOM issue is fixed.
Bug: http://b/240099508