Clean methods in capture_windows.py
This PR is just a cleaning. In a previous PR
(https://github.com/google/orbit/pull/4003) we create some methods in
a common base class that were used for testing the user-setting of auto
frame track and in this PR we are cleaning the memory e2e tests.
In addition we are also printing better log messages for all the user
settings and reusing the base class methods in Capture().

Bugs: http://b/186098691 + discussions of https://github.com/google/orbit/pull/4003
Test: Manually run memory E2E tests and in process of running all of
them in guitar.