Don't show errors related to AutoFrameTrack (#4020)
* Don't show errors related to AutoFrameTrack

In this PR we are not showing in the UI the errors from the auto added
frame track. These errors are appearing when the user doesn't wait for
the needed modules to be loaded automatically before taking a capture.
For example, an error was making the capture loading test to fail
(http://b/240382344): http://screenshot/3ztNqiMp4Mm5xwV.

To solve the issue we are using directly LoadPresetModule to load the
expected module (the shipped auto-preset has only one frame track and
therefore just one module). In case that the user modifies the file, the
behavior is unexpected and these errors might appear, and we are logging
an error to explain that. I found this solution much cleaner than using
LoadPresetModule for each module and logging a non-useful error for both
cases. Open to suggestions.

Bug: http://b/240402941
Test:
1- Manually start a capture before modules were loaded. No errors.
2- Wait modules to be loaded and take another capture. The frame track
appears automatically.

* Fixup: Reuse implementation

* Fixup: return + }

* Fixup: Add a comment for LoadPresetModule