Add default frame track when settings is turned on (#4029)
* Add default frame track when settings is turned on

In this PR we are additionally adding a the default frame track when a
user set the settings  on. This was discussed in the doc and with
Ronald and it's just to cover the case about a user who previously set
the option off and wanted to set it on. In that case it would be nice to
have the default frame track as soon as possible.

Bug: http://b/240275500

Test: Manually play with the option.

* Fixup: Capitalize If