Use "uint8_t*" instead of "char*" to represent stack data
We used char* to represent stack data that we copied via
perf_event_open. Now a "const char*" is a C-string and
has a very special semantic, while a uint8_t* reflects more
the fact that we are only looking at bytes.

An effect of this, was that GTest was trying to pretty print
our char* for the testing. However, the data to which the
pointer was pointing to was already freed, and the tests
should just compare the raw pointer.

We fixed this by using "uint8_t*" for stack data that we
collect with perf_event_open.

Test: Unit tests + Take a capture and validate callstacks
Bug: http://b/239510246
Bug: http://b/239510776