Fix crash in Cloud Collector (#4038)
There is a crash that seems to be caused by an unitialized enum variable in
`CloudCollectorStartStopCaptureRequestWaiter`. `stop_capture_reason_`
seems not to be set. I don't really understand why, but I also don't
understand the code.

My fix is to introduce a `kUnknown` enum value for the
`StopCaptureReason` and initialize the member with this value. That's
not great but it should at least save us from crashes.

BUG: http://b/240533627