Update samples tab on callstack inspections
This updates the samples tab to show only the samples in an inspection
when an inspection is active.  It also adds a banner (already present in
top-down, bottom-up views) to make it clear that the user is only seeing
a subset of samples and to leave the inspection.

Screenshot: https://screenshot.googleplex.com/479vLTQntLKZhX8

Bug: http://b/240112046