Add StopCaptureReason for the early upload failure (#4050)
* Add StopCaptureReason for the early upload failure

With this change, we add a StopCaptureReason for the early upload
failure. The CloudCollector will call StopCapture with this reason if
the upload fails early and no stop is called yet.

Bug: http://b/240533627

* Add IsStartRequested for CloudCollectorStartStopCaptureRequestWaiter

Add IsStartRequested such that the cloud collector will only call
StopCapture if start is requested. Calling StopCapture before
StartCapture will be considered as error case.

Bug: http://b/240533627