Refactor how ScopeId->FunctionInfo is found
Move scope_id_to_function_info_ into ScopeIdProvider; all of the
required fields for function_info are already there and shouldn't need
to be recalculated. GetAllProvidedScopeIds should be a superset of
instrumented_functions in CaptureData so we aren't losing any.

This will also help future PRs as my goal here is to simplify where
LiveTab gets its data so it can act on a smaller data set rather than
all of CaptureData.

Bug: http://b/240112049