Move NextPixel free function to the interface
GetNextPixelBoundaryTimeNs free function will be needed for any
implementation of TimerDataInterface (GetTimersAtDepthDiscretized will
use it to get the boundary of the pixels), so in this PR we are moving
this function to the common interface.

I also considered creating an utility file, but I think that it makes
sense to have it in TimerDataInterface.