Fix bug in Timertrack::GetYFromDepth subclasses
While implementing an optimization in SchedulerTrack I found a strange
bug while using GetYFromDepth(). This method was not retuning the
expected Y.

The solution is just override this method instead of GetYFromTimer,
since GetYFromTimer is just calling GetYFromDepth. Actually
GetYFromTimer exists just because timers have variable size in
FrameTrack, which makes things over-complicated and it's symptom of
a bad design. FrameTrack shouldn't inherit from TimerTrack (and maybe
TimerTrack should be a way smaller class or even don't exist -
http://b/203181055)

Anyway, in this PR we are just fixing this for Scheduler and
GpuDebugMarkerTrack. GpuSubmissionTrack has a really weird logic about
how to calculate depth that I don't want to touch. We spoke with Florian
several months ago and I will leave it as it is for now.

Test: Load a capture, everything works as expected.