Fix glitch in ThreadTimers because intersection
In this PR we are fixing a glitch related to intersection of timers in
ThreadTrack. It's not very visible because our way to render timers but
it will be after extending timers to the border of the pixels.

Basically we are using an alpha to render timer in even depths, to
differentiate of the odd depths, but alpha and intersection doesn't work
as expected and we ended having a lighter color if 2 timers has
intersection in the same pixel.

In this PR we are computing the new color as if we have an alpha but
without using the alpha-value. Therefore, intersection wouldn't be an
issue.

Before: http://screenshot/55KLpRYwkXnRfsY. The lightest part should be
the leftmost pixel even with anti-aliasing.
After: http://screenshot/B47EKrWEodgNEeC

Bug: http://b/242973688.
Test: Load different captures. Play with them.