Add ScopeId value check in LiveFunctionsDataView
The minidump and the callstack available for the recent crash could not
shed light on the nature of the issue. To that end it was decided
to add an explicit check upon adding a ScopeId to
`LiveFunctionsDataView::idices_`. That's it, we verify the id to be
in the range of the ids issued by `ScopeIdProvider`.

Test: Unit, Manual
Bug: http://b/243122633