Make DeduceDebugSymbolMissingSizes available for all ObjectFiles
DeduceDebugSymbolMissingSizes is a free function in CoffFile
that computes the missing size information of some records
based on the address of the next symbol.

We will need to do exactly the same for Pdb files when
when reading the public symbol stream with LLVM
(DIA seem to do this internally). So in order to share
code, we move this function (and the dependands)
up to ObjectFile.

Test: Compile