Fix crashes when multiple Vulkan layers are loaded (#4113)
They were two conception problems when in vkCreateDevice
that may lead to a crash when another Vulkan layer gets
loaded.

1. Calling vkGetInstanceProcAddr with VK_NULL_HANDLE
In order to call vkCreateDevice down the chain, we
retrieve the function pointer using the vkGetInstanceProcAddr
function for vkCreateDevice.
As we did not care about the instance, we pass in "VK_NULL_HANDLE",
as described in the Vulkan layer tutorial and almost all other
Vulkan layer. However, vkCreateDevice is not a global function,
but an instance function, so per API, this is undefined and the
subsequent call to the function pointer might
fail. While most ICDs and layers will return the right function
pointer anyways, Renderdoc returns null in that case.

So we do the right thing, and call vkGetInstanceProcAddr with
the actual instance.

2. `u.pLayerInfo->pfnNextGetInstanceProcAddr` is not always
   returning the function pointer down the chain.
For most instance functions, like "vkEnumerateDeviceExtensionProperties"
the Vulkan loader assumes that after the "vkCreateInstance" call
you are not doing another lookup using the pfnNextGetInstanceProcAddr
pointer, but instead create a dispatch table.
If layer A is before our Vulkan layer, we might end up in
getting the function pointer of layer A and fail.
You can only get the proper address for vkGetInstanceProcAddr,
vk_layerGetPhysicalDeviceProcAddr, vkCreateInstance, vkCreateDevice
and some debug util functions.

So we use the cached dispatch table instead.
Some more context can be found in http://b/243079253.

Test: load the Vulkan layer together with RenderDoc on a couple
      of games.

Bug: http://b/243079253