Add SymbolLoadingOutcome
SymbolLoadingOutcome expresses the result of a symbol loading task that
can fail with an error, get canceled, result in a "NotFound" outcome, or
success. This will be used as a return type to various symbol loading
operations. The `SuccessType` outcome contains a file path and 2 fields
that contain meta data about the search. These will be used for metrics
(compare SymbolLoadingMetric and/or orbit_log_event.proto)

Test: Unit Test
Bug: http://b/231455031