Parese PDB's public symbols using LLVM
In a PDB file, public symbols are not a real subset of
the module info stream's symbols. There are symbols for
functions (like RtUserThreadStart) that are only present
in the public symbols.

This change adds support for reading public symbols
using LLVM.

Unlike DIA, LLVM is not extrapolating the size information
for the public symbols. Therefore, we try to deduce
this information from the distance to the next
symbol (similar to what we do in the CoffFile).
For the last symbol, we try to find the size information
in the section contribution stream.

Test: Run test + Check on a couple of Windows dlls.
Bug: http://b/221968448