Add SymbolFinder and move symbol_files_currently_dwonloading_
With this change, we created the `SymbolFinder` class and moved
`symbol_files_currently_downloading` from `OrbitApp` to `SymbolFinder`.

This is the first PR for the symbol loading refactoring. All the logic
for retrieving a local symbol file path for a given module will be moved
from `OrbitApp` to `SymbolFinder` gradually.

Bug: http://b/243520787